Rename sap module to sap_client
